util/hub_eval/pyproject.toml
[tool.poetry]
name = "hub_eval"
version = "0.1.0"
description = ""
authors = ["Vicente Herrera <vicenteherrera@vicenteherrera.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
